Implemented all the basic commands as cogs
Separated deckdiff into a cog

Separated deckdiff into a cog

Moved all the commands to cogs

Moved all the commands to cogs

Updated the requirements

Re-implemented the basic card fetcher functionality as a cog

Rewrote the channel cleaner, finished the fetcher implementation

Re-implemented the price command

Re-implemented the channel cleaner
