+++ run.py
@@ -5,5 +5,6 @@ from src import bot
 
 logging.basicConfig(level=logging.INFO)
 token = os.environ['BOLAS_SECRET_TOKEN']
+
 bot = bot.Bolas()
 bot.run(token,reconnect=True)
